Add more comments and change theme
Added more comments to explain what is going on and give some props where
needed and changed theme to molokai again.

I also removed the auch chd dir to current files dir, because it was just
frustrating me and I'll stick to the old fashioned way for now.

I also the background to dark, because it makes things pretty in the terminal
and that is where I mostly use vim.